requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
pydantic==2.6.1
pydantic_core==2.16.2
python-multipart==0.0.9
pytz==2024.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
